+++ src/libmach/5.c
@@ -35,7 +35,7 @@
 #include "ureg_arm.h"
 #include <mach.h>
 
-#define	REGOFF(x)	(ulong) (&((struct Ureg *) 0)->x)
+#define	REGOFF(x)	(uintptr) (&((struct Ureg *) 0)->x)
 
 #define SP		REGOFF(r13)
 #define PC		REGOFF(pc)
